Setup Mastery component Core data stuctures
Sets up the Mastery component Core data structures.

The Mastery component will be Quiz users to twst their mastery of the
category.

To prove Mastery the user will need to answer a certain number of
questions correctly in a row.

Each catgory's Question will be gnerated by a Template so that there is
a mechanism to generate questions dynamically.

This touches upon how to organise data structures in a Functional
environment.

A stand out excerpt is:

> Next we introduced the way functional programmers shape data, preferring
> many versions of a value over time rather than continuously mutating a single
> value.