Build out the Mastery component Functional Core
Builds out the Mastery component Functional Core.

There's a lot going on here but the main piece is the Quiz module.

Most interactions with a quiz appear to be through that module and it
handles the "operations" that can be performed on a single quiz:

- Create a Quiz
- Add a Template to a Quiz
- Select a Question to ask a user
- Answer the selected Question

It does a lot of stuff and it feels like the "business" part of this
should be pulled out of the Quiz struct to lie in a Context/Business
logic module.

I added some really basic tests that aren't great and build up test data
using the functions under test themselves but they helped with
understanding what was going on in the code and how things pieced
together more than the book's IEx example did.